07-21

+++ controller/post.js
@@ -2,6 +2,8 @@
 
 const Sequelize = require('sequelize');
 const lodash = require('lodash');
+const util = require('util');
+const Promise = require('bluebird');
 
 const mysql = require('../lib/mysql');
 const tools = require('../lib/tools')
@@ -13,47 +15,41 @@ const logger = global.logger;
 
 const { checkpara_int, checkpara_str } = tools;
 
-module.exports = { getPage, publish, getInfo, del };
+module.exports = {
+  getPage, publish, getInfo, deletePost, getDetail, getReplyPage,
+  sendReply, deleteReply, getTipPage
+};
 
 /**
  * 获取帖子列表
  * @param {string} key 关键字 
  * @param {int} page 页码
  * @param {int} size 页面大小
- * @param {int} isme 是否查看本人帖子 1-是
+ * @param {int} me 是否查看本人帖子 1-是
  */
 async function getPage({ params, user, endfor }) {
-  const { key, page = 1, size = 20, isme = 0 } = params;
-  if (!checkpara_int(page) || !checkpara_int(size)) 
+  const { key, page = 1, size = 20, me = 0 } = params;
+  if (!checkpara_int(page) || !checkpara_int(size))
     return endfor(ET.缺少必须参数);
 
-  const where = { State: Post.ESTATE.启用 };
-  if (isme) {
-    where.UserId = user.Id;
-    where.State = {
-      $in: [Post.ESTATE.启用, Post.ESTATE.草稿]
-    }
-  }
-  if (key) where.Title = { $like: `%${key}%` };
+  let sql = 'select %s from user,post left join post_dept on post.Id = post_dept.post_id' +
+    ` where %s order by is_top desc,post.id desc limit ${(page - 1) * size},${size}`;
 
-  const results = await Post.findAndCountAll({
-    attributes: ['Id', 'Title', 'UserId', 'Count', 'State', 'CreateTime', 'IsTop'],
-    where,
-    raw: true,
-    include: {
-      model: 'PostDept',
-      where: {
-        $or: [{ IsPublic: 0 }, { DeptId: user.DeptId }]   //fix me
-      },
-      require: true
-    },
-    order: [['IsTop', 'DESC'], ['CreateTime', 'DESC']],
-    limit: size,
-    offset: (page - 1) * size,
-  }).catch(err => logger.error(`post.getpage查询失败,${err.message}`));
+  let where = `user.id = post.user_id and (post.is_public = 1 or post_dept.dept_id = ${user.DeptId})`;
+  if (me == 1) {
+    where += ` and post.user_id = ${user.Id} and state in (${Post.ESTATE.正常}, ${Post.ESTATE.草稿})`;
+  } else {
+    where += ` and post.state = ${Post.ESTATE.正常}`;
+  }
+  if (key) where += ` and title like %${key}%`;
+
+  const fields = ['user.name, title, is_top, post.count, post.create_time', 'count(*) as count'];
+  const results = await Promise.map(fields, field =>
+    mysql.query(util.format(sql, field, where), { type: mysql.QueryTypes.SELECT })
+  ).catch(err => logger.error(`post.getPage查询post失败,${err.message}`));
 
   if (!results) return endfor(ET.数据异常);
-  return endfor(ET.成功, { count: results.count, items: result.rows });
+  return endfor(ET.成功, { count: results[1][0].count, items: results[0] });
 }
 
 /**
@@ -62,7 +58,7 @@ async function getPage({ params, user, endfor }) {
  * @param {int} is_public 帖子是否公开可见
  */
 async function publish({ params, user, endfor }) {
-  const { id, title, state = Post.ESTATE.启用, is_public = 1, is_attach = 0, 
+  const { id, title, state = Post.ESTATE.正常, is_public = 1, is_attach = 0,
     dept_id, content } = params;
 
   if (!(title && content)) return endfor(ET.缺少必须参数);
@@ -70,8 +66,8 @@ async function publish({ params, user, endfor }) {
   if (!checkpara_str(content)) return endfor(ET.参数不合法);
   if (!checkpara_int(is_public, 0, 1)) return endfor(ET.参数不合法);
   if (!checkpara_int(is_attach, 0, 1)) return endfor(ET.参数不合法);
-  if (!checkpara_int(state, Post.ESTATE.启用, Post.ESTATE.草稿))
-  return endfor(ET.参数不合法);
+  if (!checkpara_int(state, Post.ESTATE.正常, Post.ESTATE.草稿))
+    return endfor(ET.参数不合法);
   if (is_public != 0 && !/^\d+(,\d+)*$/.test(dept_id))
     return endfor(ET.参数不合法);
   if (+is_attach === 1 && user.IsAdmin !== 1)
@@ -79,20 +75,20 @@ async function publish({ params, user, endfor }) {
 
   let post;
   if (id) {
-    post = Post.find({
-      where: { Id: id, State: Post.ESTATE.启用 }
+    post = await Post.findById(+id, {
+      where: { State: Post.ESTATE.正常 }
     }).catch(err => logger.error(`post.publish查询失败,${err.message}`));
     if (!post) return endfor(ET.记录不存在);
 
-    post = post.update({
-      Content, content, IsPublic: is_public,
+    post = await post.update({
+      Content: content, IsPublic: is_public,
       Title: title, State: state, AllowAttach: is_attach
     }).catch(err => logger.error(`post.publish更新失败,${err.message}`));
     if (!post) return endfor(ET.数据异常);
 
   } else {
-    post = post.create({
-      UserId: user.Id, Content, content, IsPublic: is_public,
+    post = await Post.create({
+      UserId: user.Id, Content: content, IsPublic: is_public,
       Title: title, State: state, AllowAttach: is_attach
     }).catch(err => logger.error(`post.publish更新失败,${err.message}`));
 
@@ -100,14 +96,15 @@ async function publish({ params, user, endfor }) {
   }
 
   //创建可见关系记录
-  if (dept_id) {
+  if (is_public == 0 && dept_id) {
     const postDepts = await mysql.transaction(() =>
       PostDept.destroy({
         where: { PostId: post.Id }
       }).then(() =>
         PostDept.bulkCreate(dept_id.split(',').map(id => {
-          return { PostId: postid, DeptId: +id }
-        })))
+          return { PostId: post.Id, DeptId: +id }
+        }))
+      )
     ).catch(err => logger.error(`post.publish增删PostDept失败,${err.message}`));
 
     if (!postDepts) return endfor(ET.数据异常);
@@ -120,13 +117,13 @@ async function publish({ params, user, endfor }) {
  * 删除帖子
  * @param {*} param0 
  */
-async function del({ params, user, endfor }) {
+async function deletePost({ params, user, endfor }) {
   const { id } = params;
   if (!id) return endfor(ET.缺少必须参数);
+  if (!checkpara_int(id)) return endfor(ET.参数不合法);
 
-  const post = await Post.find({
+  let post = await Post.findById(+id, {
     where: {
-      Id: id,
       State: {
         $not: Post.ESTATE.删除
       }
@@ -138,7 +135,7 @@ async function del({ params, user, endfor }) {
     return endfor(ET.没有权限);
 
   post = await post.update({ State: Post.ESTATE.删除 })
-    .create(err => logger.error(`post.del删除失败,${err.message}`));
+    .catch(err => logger.error(`post.del删除失败,${err.message}`));
   if (!post) return endfor(ET.数据异常);
 
   return endfor(ET.成功);
@@ -151,10 +148,11 @@ async function del({ params, user, endfor }) {
 async function getInfo({ params, endfor }) {
   const { id } = params;
   if (!id) return endfor(ET.缺少必须参数);
+  if (!checkpara_int(id)) return endfor(ET.参数不合法);
 
-  let post = await Post.findById(id, {
-    attributes: ['Title', 'UserId', 'IsAllowAttach', 'Content',
-      'Count', 'CreateTime'],
+  let post = await Post.findById(+id, {
+    attributes: ['Id', 'Title', 'UserId', 'AllowAttach', 'Content',
+      'Count', 'create_time'],
     where: {
       State: {
         $not: Post.ESTATE.删除
@@ -164,14 +162,14 @@ async function getInfo({ params, endfor }) {
       model: User,
       attributes: ['Number', 'Name', 'Avater', 'State'],
       require: true
-    },
-    raw: true
+    }
   }).catch(err => logger.error(`post.getInfo查询失败,${err.message}`));
   if (!post) return endfor(ET.记录不存在);
 
   post = await post.update({
     Count: mysql.literal('`count` + 1')
-  }).catch(err => logger.error(`post.getInfo更新失败,${err.message}`));
+  }).then(post => post.reload())
+    .catch(err => logger.error(`post.getInfo更新失败,${err.message}`));
   if (!post) return endfor(ET.数据异常);
 
   return endfor(ET.成功, post);
@@ -184,17 +182,17 @@ async function getInfo({ params, endfor }) {
 async function getDetail({ params, user, endfor }) {
   const { id } = params;
   if (!id) return endfor(ET.缺少必须参数);
+  if (!checkpara_int(id)) return endfor(ET.参数不合法);
 
-  const post = await Post.find({
+  const post = await Post.findById(+id, {
     where: {
-      Id: +id,
       State: {
         $not: Post.ESTATE.删除
       }
     }
   }).catch(err => logger.error(`post.getdetail查询失败,${err.message}`));
   if (!post) return endfor(ET.记录不存在);
-  if (post.Id !== user.Id)
+  if (post.UserId !== user.Id)
     return endfor(ET.没有权限);
 
   //帖子为不公开，获取允许查看的部门
@@ -219,35 +217,55 @@ async function getDetail({ params, user, endfor }) {
  * 获取评论列表
  * @param {*} param0 
  */
-async function getReplys({ params, endfor }) {
+async function getReplyPage({ params, endfor }) {
   const { id, page = 1, size = 10 } = params;
   if (!id) return endfor(ET.缺少必须参数);
+  if (!checkpara_int(id)) return endfor(ET.参数不合法);
+  if (!checkpara_int(page)) return endfor(ET.参数不合法);
+  if (!checkpara_int(size)) return endfor(ET.参数不合法);
 
-  const post = await Post.findById(+id, {
-    where: { State: Post.ESTATE.启用 }
+  const post = await Post.findOne({
+    where: {
+      Id: +id,
+      State: Post.ESTATE.正常
+    }
   }).catch(err => logger.error(`post.getReplys查询post失败,${err.message}`));
   if (!post) return endfor(ET.记录不存在);
 
   const replys = await Reply.findAndCount({
     where: {
-      PostId: id,
+      PostId: +id,
       State: 0,
       Type: Reply.ETYPE.一级回复
     },
     limit: size,
     offset: (page - 1) * size,
     include: {
-      model: Reply,
-      where: { State: 0 },
-      include: {
-        model: User,
-        attributes: ['Id', 'Name', 'Avater']
-      },
+      model: User,
+      attributes: ['Id', 'Name', 'Avater', 'State']
     },
     order: [['Id', 'DESC']],
     raw: true,
-  }).catch(err => logger.error(`post.getReplys查询`));
+  }).catch(err => logger.error(`post.getReplys查询,${err.message}`));
   if (!replys) return endfor(ET.数据异常);
+
+  const replyIds = replys.rows.map(reply => reply.Id);
+  const subReplys = await Reply.findAll({
+    where: {
+      Reply_Id1: { $in: replyIds },
+    },
+    include: {
+      model: User,
+      attributes: ['Id', 'Name', 'Avater', 'State']
+    },
+  }).catch(err => logger.error(`post.getReplyPage查询subReply失败,${err.message}`));
+  if (!subReplys) return endfor(ET.数据异常);
+
+  replys.rows.forEach(reply => {
+    reply.subReplys = subReplys.filter((subReply) => {
+      return subReply.ReplyId1 === reply.Id;
+    })
+  });
 
   return endfor(ET.成功, { items: replys.rows, count: replys.count });
 }
@@ -257,54 +275,60 @@ async function getReplys({ params, endfor }) {
  * @param {*} param0 
  */
 async function sendReply({ params, user, endfor }) {
-  const { id, replyId1, replyId2 } = params;
-  if (!id) return endfor(ET.缺少必须参数);
+  const { id, reply_id1, reply_id2, content } = params;
+  if (!id || !content) return endfor(ET.缺少必须参数);
+  if (!checkpara_int(reply_id1)) return endfor(ET.参数不合法);
+  if (!checkpara_int(reply_id2)) return endfor(ET.参数不合法);
+  if (!checkpara_str(content, 1)) return endfor(ET.参数不合法);
 
-  const post = await Post.findById(+id, {
-    where: { State: Post.ESTATE.启用 }
+  const post = await Post.findOne({
+    where: {
+      Id: id,
+      State: Post.ESTATE.正常
+    }
   }).catch(err => logger.error(`post.setReply查询post失败,${err.message}`));
   if (!post) return endfor(ET.记录不存在);
 
   const value = {
-    PostId: id, UserId: user.Id, Type: type
+    PostId: id, UserId: user.Id, Content: content
   };
   value.Type = Reply.ETYPE.一级回复;
   let beUserId = post.UserId; //被回复人id
-  if (replyId1) {
-    const reply = await Reply.find({
+  if (reply_id1) {
+    const reply = await Reply.findOne({
       where: {
-        ReplyId1: +replyId1,
+        id: reply_id1,
         State: 0
       }
     }).catch(err => logger.error(`post.setReply查询reply1失败,${err.message}`));
     if (!reply) return endfor(ET.记录不存在);
 
-    value.ReplyId1 = replyId1;
+    value.ReplyId1 = reply_id1;
     value.Type = Reply.ETYPE.二级回复;
     beUserId = reply.UserId;
   }
 
-  if (replyId2) {
-    const reply = await Reply.find({
+  if (reply_id2) {
+    const reply = await Reply.findOne({
       where: {
-        ReplyId2: +replyId2,
+        id: +reply_id2,
         State: 0
       }
     }).catch(err => logger.error(`post.setReply查询reply2失败,${err.message}`));
     if (!reply) return endfor(ET.记录不存在);
 
-    value.ReplyId2 = replyId2;
+    value.ReplyId2 = reply_id2;
     value.Type = Reply.ETYPE.三级回复;
     beUserId = reply.UserId;
   }
 
   const tip = await mysql.transaction(() =>
-    Reply.create(value).then(() => {
-      return Tip.create({
+    Reply.create(value).then((reply) =>
+      Tip.create({
         PostId: id, UserId: beUserId,
         ReplyUserId: user.Id, ReplyId: reply.Id
       })
-    })
+    )
   ).catch(err => logger.error(`post.setReply创建失败,${err.message}`));
   if (!tip) return endfor(ET.数据异常);
 
@@ -318,10 +342,11 @@ async function sendReply({ params, user, endfor }) {
 async function deleteReply({ params, user, endfor }) {
   const { id } = params;
   if (!id) return endfor(ET.缺少必须参数);
-  if (!isNaN(id)) return endfor(ET.参数不合法);
+  if (!checkpara_int(id)) return endfor(ET.参数不合法);
 
-  const reply = Reply.findById(id, {
+  let reply = await Reply.findOne({
     where: {
+      Id: +id,
       State: 0
     }
   }).catch(err => logger.error(`post.deleteReply查询失败,${err.message}`));
@@ -329,13 +354,10 @@ async function deleteReply({ params, user, endfor }) {
 
   if (reply.UserId !== user.Id) return endfor(ET.没有权限);
 
-  const count = await Reply.destroy({
-    where: {
-      Id: id,
-      State: 0
-    }
-  }).catch(err => logger.error(`post.deleteReply删除失败,${err.message}`));
-  if (count !== 1) return endfor(ET.数据异常);
+  reply = await reply.update({
+    State: 1
+  }).catch(err => logger.error(`post.deleteReply更新失败,${err.message}`));
+  if (!reply) return endfor(ET.数据异常);
 
   return endfor(ET.成功);
 }
@@ -346,6 +368,9 @@ async function deleteReply({ params, user, endfor }) {
  */
 async function getTipPage({ params, user, endfor }) {
   const { page = 1, size = 20 } = params;
+  if (!checkpara_int(page)) return endfor(ET.参数不合法);
+  if (!checkpara_int(size)) return endfor(ET.参数不合法);
+
   const result = await Tip.findAndCount({
     where: {
       UserId: user.Id,
@@ -353,12 +378,12 @@ async function getTipPage({ params, user, endfor }) {
     order: [['Id', 'DESC']],
     include: [
       {
-        model: 'Post',
+        model: Post,
         attributes: ['State', 'Id', 'Title'],
         require: true
       },
       {
-        model: 'User',
+        model: User,
         attributes: ['Id', 'Name'],
         require: true
       }
